requirements.txt
nmap==0.0.1
ping3==4.0.4
python-dotenv==0.21.1
python-nmap==0.7.1
python3-nmap==1.6.0
simplejson==3.18.3
